chief-of-staff/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0
alembic>=1.13.0
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
anthropic>=0.40.0
cryptography>=42.0
httpx>=0.27.0
google-auth>=2.28.0
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0
arq>=0.26.0
redis>=5.0
asyncpg>=0.29.0
aioapns>=3.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.7
pytest>=8.0
email-validator>=2.1.0
